fixed JWT generation & do not load AWS cedrentials in AWS

+++ course-02/exercises/udacity-c2-restapi/src/controllers/v0/users/routes/auth.router.ts
@@ -27,7 +27,7 @@ async function comparePasswords(plainTextPassword: string, hash: string): Promis
 
 function generateJWT(user: User): string {
     //@TODO Use jwt to create a new JWT Payload containing
-    return jwt.sign(user, config.jwt.secret);
+    return jwt.sign(user.short(), config.jwt.secret);
 }
 
 export function requireAuth(req: Request, res: Response, next: NextFunction) {
@@ -124,7 +124,6 @@ router.post('/', async (req: Request, res: Response) => {
     } catch (e) {
         throw e;
     }
-
     // Generate JWT
     const jwt = generateJWT(savedUser);
